UserMetadata isn't used either, but removing ProfileBadge to avoid confusion in future
